% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbh_brass_qx.R
\name{sbh_brass_qx}
\alias{sbh_brass_qx}
\title{Compute q(x) via Brass (Trussell Variant) method using SBH information}
\usage{
sbh_brass_qx(region_dataframe, i_brass, family)
}
\arguments{
\item{i_brass}{i corresponding to the q(x) being calculated. for example, i = 1 (ages 15-19) corresponds to x = 1, i = 2 (ages 15-19) corresponds to x = 2, etc. - this should be a single integer ranging from 1-7 (inclusive).}

\item{family}{string representing family of Coale-Demeny life tables to be used (options: East, West, North, South).}

\item{district_dataframe}{a dataframe for a single region/area with SBH information for 5.yr age bands & both sexes from 15-19 up to 45-49, and a corresponding index column (i) for the age groups... columns should be named: region, i, num_women, total_both, dead_both. all less region should be numeric. the dataframe should be 7 rows.}
}
\value{
the region name from the dataframe, the computed q(x), and the multiplier k(x) from the associated formula
}
\description{
Compute q(x) via Brass (Trussell Variant) method using SBH information
}
