% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbh_brass_ti.R
\name{sbh_brass_ti}
\alias{sbh_brass_ti}
\title{Compute t(i) for calculated q(x)'s from the Brass (Trussell Variant) method using SBH information}
\usage{
sbh_brass_ti(region_dataframe, i_brass, family, census_year)
}
\arguments{
\item{region_dataframe}{a dataframe for a single region/area with SBH information for 5.yr age bands & both sexes from 15-19 up to 45-49, and a corresponding index column (i) for the age groups... columns should be named: region, i, num_women, total_both, dead_both. all less region should be numeric. the dataframe should be 7 rows.}

\item{i_brass}{i corresponding to the q(x) of interest for the reference year. for example, i = 1 (ages 15-19) corresponds to x = 1, i = 2 (ages 15-19) corresponds to x = 2, etc. - this should be a single integer ranging from 1-7 (inclusive).}

\item{family}{string representing family of Coale-Demeny life tables to be used (options: East, West, North, South).}

\item{census_year}{single integer representing the year of the census being used}
}
\value{
the reference year for a given i corresponding to the q(x) of interest
}
\description{
Compute t(i) for calculated q(x)'s from the Brass (Trussell Variant) method using SBH information
}
